refactor: split the yt-dlp engine into modules behind an unchanged barrel
